Simplify DefaultVisualization/Item, make it a simple component
VisualizationItem/Item is the connected component, whatever is required by
DefaultVisualizationItem and other item types like ChartPlugin is simply
passed as props.

Also the computation for the height is done in one place and passed in
the style prop to the children.
